+++ src/__tests__/__fixtures__/validationFailure/src/validateInvocation.ts
@@ -1,0 +1,5 @@
+import { IntegrationValidationError } from '../../../../framework';
+
+export default function validateInvocation() {
+  throw new IntegrationValidationError('Failed to access provider api');
+}